organize: classes by context
