+++ brut.apktool/apktool-lib/src/main/java/brut/androlib/res/AaptInvoker.java
@@ -14,8 +14,9 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package brut.androlib;
+package brut.androlib.res;
 
+import brut.androlib.Config;
 import brut.androlib.exceptions.AndrolibException;
 import brut.androlib.meta.ApkInfo;
 import brut.common.BrutException;
