+++ brut.apktool/apktool-lib/src/main/java/brut/androlib/res/table/value/ResEnum.java
@@ -72,7 +72,7 @@ public class ResEnum extends ResAttribute {
     }
 
     @Override
-    public void serializeItemsToResValuesXml(XmlSerializer serial, ResEntry entry)
+    public void serializeSymbolsToValuesXml(XmlSerializer serial, ResEntry entry)
             throws AndrolibException, IOException {
         Config config = entry.getPackage().getTable().getConfig();
         boolean removeUnresolved = config.getDecodeResolve() == Config.DecodeResolve.REMOVE;
