+++ brut.apktool/apktool-lib/src/test/java/brut/androlib/res/AaptVersionTest.java
@@ -14,8 +14,9 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package brut.androlib;
+package brut.androlib.res;
 
+import brut.androlib.BaseTest;
 import brut.androlib.exceptions.AndrolibException;
 
 import org.junit.*;
